Move account providers to a submodule
